Add styling for fake Bootstrap buttons
The .fake class disables all styling related to pseudo-classes when used on an element styled as a Bootstrap button. It can be used to make a non-interactive element look (but not behave) like a Bootstrap button. It is intended to be used in a button group to act as a label or subject for other buttons in the group.